Packaging and configuring system improved. Now making use of config.h. Man page no longer gzipped and now automaticaly versioned. Fixed two warnings in main.cpp caused by dev.h. Printhelp call no longer has an argument.

+++ nload/dev.h
@@ -29,8 +29,8 @@
 #include "graph.h"
 #include "status.h"
 
-const long STANDARD_BAR_MAX_IN = 10240;
-const long STANDARD_BAR_MAX_OUT = 10240;
+const int STANDARD_BAR_MAX_IN = 10240;
+const int STANDARD_BAR_MAX_OUT = 10240;
 const bool STANDARD_SHOW_GRAPHS = true;
 const int STANDARD_SLEEP_INTERVAL = 500;
 char* const STANDARD_NETWORK_DEVICE = "eth0";
